refactor: feature-based packaging
