+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/ebook/repository/EbookJpaRepository.java
@@ -1,9 +1,8 @@
-package qwerty.chaekit.domain.ebook;
+package qwerty.chaekit.domain.ebook.repository;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import qwerty.chaekit.domain.ebook.Ebook;
 
 @Repository
 public interface EbookJpaRepository extends JpaRepository<Ebook, Long> {
