+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/ebook/repository/EbookRepository.java
@@ -1,8 +1,9 @@
-package qwerty.chaekit.domain.ebook;
+package qwerty.chaekit.domain.ebook.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
+import qwerty.chaekit.domain.ebook.Ebook;
 
 import java.util.Optional;
