+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/group/activity/discussion/Discussion.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import qwerty.chaekit.domain.BaseEntity;
 import qwerty.chaekit.domain.group.activity.Activity;
+import qwerty.chaekit.domain.group.activity.discussion.comment.DiscussionComment;
 import qwerty.chaekit.domain.member.user.UserProfile;
 
 import java.util.ArrayList;
