+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/group/activity/discussion/comment/repository/DiscussionCommentRepositoryImpl.java
@@ -1,10 +1,10 @@
-package qwerty.chaekit.domain.group.activity.discussion.comment;
+package qwerty.chaekit.domain.group.activity.discussion.comment.repository;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import qwerty.chaekit.domain.group.activity.discussion.DiscussionComment;
 import qwerty.chaekit.domain.group.activity.discussion.QDiscussionComment;
+import qwerty.chaekit.domain.group.activity.discussion.comment.DiscussionComment;
 
 import java.util.List;
 import java.util.Map;
