+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/group/groupmember/GroupMemberRepository.java
@@ -1,0 +1,12 @@
+package qwerty.chaekit.domain.group.groupmember;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import qwerty.chaekit.domain.group.ReadingGroup;
+
+@Repository
+public interface GroupMemberRepository extends JpaRepository<GroupMember, Long> {
+    Page<GroupMember> findByReadingGroupAndAcceptedFalse(ReadingGroup group, Pageable pageable);
+}
