+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/group/grouptag/GroupTagRepository.java
@@ -1,0 +1,7 @@
+package qwerty.chaekit.domain.group.grouptag;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GroupTagRepository extends JpaRepository<GroupTag, Long> {
+
+}
