--- chaekit-spring/src/main/java/qwerty/chaekit/domain/group/tag/GroupTagRepository.java
@@ -1,7 +1,0 @@
-package qwerty.chaekit.domain.group.tag;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GroupTagRepository extends JpaRepository<GroupTag, Long> {
-
-}
