+++ chaekit-spring/src/main/java/qwerty/chaekit/dto/group/GroupFetchResponse.java
@@ -1,7 +1,7 @@
 package qwerty.chaekit.dto.group;
 
 import lombok.Builder;
-import qwerty.chaekit.domain.group.tag.GroupTag;
+import qwerty.chaekit.domain.group.grouptag.GroupTag;
 import qwerty.chaekit.domain.group.ReadingGroup;
 import qwerty.chaekit.dto.group.enums.MyMemberShipStatus;
