+++ chaekit-spring/src/main/java/qwerty/chaekit/dto/group/GroupPostResponse.java
@@ -2,7 +2,7 @@ package qwerty.chaekit.dto.group;
 
 import lombok.Builder;
 import qwerty.chaekit.domain.group.ReadingGroup;
-import qwerty.chaekit.domain.group.tag.GroupTag;
+import qwerty.chaekit.domain.group.grouptag.GroupTag;
 
 import java.util.List;
