+++ chaekit-spring/src/main/java/qwerty/chaekit/service/group/ActivityService.java
@@ -5,12 +5,11 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import qwerty.chaekit.domain.ebook.EbookJpaRepository;
-import qwerty.chaekit.domain.ebook.EbookRepository;
-import qwerty.chaekit.domain.group.GroupRepository;
+import qwerty.chaekit.domain.ebook.repository.EbookRepository;
+import qwerty.chaekit.domain.group.repository.GroupRepository;
 import qwerty.chaekit.domain.group.ReadingGroup;
 import qwerty.chaekit.domain.group.activity.Activity;
-import qwerty.chaekit.domain.group.activity.ActivityRepository;
+import qwerty.chaekit.domain.group.activity.repository.ActivityRepository;
 import qwerty.chaekit.domain.member.user.UserProfileRepository;
 import qwerty.chaekit.dto.group.activity.ActivityFetchResponse;
 import qwerty.chaekit.dto.group.activity.ActivityPatchRequest;
