+++ chaekit-spring/src/main/java/qwerty/chaekit/service/group/GroupService.java
@@ -5,9 +5,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import qwerty.chaekit.domain.group.GroupMember;
-import qwerty.chaekit.domain.group.GroupMemberRepository;
-import qwerty.chaekit.domain.group.GroupRepository;
+import qwerty.chaekit.domain.group.groupmember.GroupMember;
+import qwerty.chaekit.domain.group.groupmember.GroupMemberRepository;
+import qwerty.chaekit.domain.group.repository.GroupRepository;
 import qwerty.chaekit.domain.group.ReadingGroup;
 import qwerty.chaekit.domain.member.user.UserProfile;
 import qwerty.chaekit.domain.member.user.UserProfileRepository;
