+++ chaekit-spring/src/test/java/qwerty/chaekit/util/TestFixtureFactory.java
@@ -3,8 +3,8 @@ package qwerty.chaekit.util;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import qwerty.chaekit.domain.ebook.Ebook;
-import qwerty.chaekit.domain.ebook.EbookRepository;
-import qwerty.chaekit.domain.group.GroupRepository;
+import qwerty.chaekit.domain.ebook.repository.EbookRepository;
+import qwerty.chaekit.domain.group.repository.GroupRepository;
 import qwerty.chaekit.domain.group.ReadingGroup;
 import qwerty.chaekit.domain.member.Member;
 import qwerty.chaekit.domain.member.MemberRepository;
